move wallet code
